# pkg/web/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for querying information about Assertoor tests",
        "title": "Assertoor API Documentation",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/test_run/{runId}/cancel": {
            "post": {
                "description": "Returns the test/run id \u0026 status of the cancelled test.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TestRun"
                ],
                "summary": "Cancel test run by test ID",
                "operationId": "postTestRunCancel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the test run to cancel",
                        "name": "runId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Test cancellation options",
                        "name": "cancelOptions",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web_api.PostTestRunCancelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web_api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web_api.PostTestRunCancelResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/web_api.Response"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "$ref": "#/definitions/web_api.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/tests/register": {
            "post": {
                "description": "Returns the test id and name of the added test.",
                "consumes": [
                    "application/json",
                    "application/yaml"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Register new test via yaml configuration",
                "operationId": "postTestsRegister",
                "parameters": [
                    {
                        "description": "Test configuration (json or yaml)",
                        "name": "testConfig",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web_api.PostTestsRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web_api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web_api.PostTestsRegisterResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/web_api.Response"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "$ref": "#/definitions/web_api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/tests/register_external": {
            "post": {
                "description": "Returns the test id and name of the added test.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Register new test via external test configuration",
                "operationId": "postTestsRegisterExternal",
                "parameters": [
                    {
                        "description": "Test configuration (json or yaml)",
                        "name": "externalTestConfig",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web_api.PostTestsRegisterExternalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web_api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web_api.PostTestsRegisterExternalResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/web_api.Response"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "$ref": "#/definitions/web_api.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "helper.RawMessage": {
            "type": "object"
        },
        "types.TestSchedule": {
            "type": "object",
            "properties": {
                "cron": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "skipQueue": {
                    "type": "boolean"
                },
                "startup": {
                    "type": "boolean"
                }
            }
        },
...
        "web_api.PostTestRunCancelRequest": {
            "type": "object",
            "properties": {
                "skip_cleanup": {
                    "type": "boolean"
                },
                "test_id": {
                    "type": "string"
                }
            }
        },
        "web_api.PostTestRunCancelResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "run_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "test_id": {
                    "type": "string"
                }
            }
        },
...
        "web_api.PostTestsRegisterExternalRequest": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object",
                    "additionalProperties": true
                },
                "configVars": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "file": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "schedule": {
                    "$ref": "#/definitions/types.TestSchedule"
                },
                "timeout": {
                    "type": "integer"
                }
            }
        },
        "web_api.PostTestsRegisterExternalResponse": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "name": {
                    "type": "string"
                },
                "test_id": {
                    "type": "string"
                }
            }
        },
        "web_api.PostTestsRegisterRequest": {
            "type": "object",
            "properties": {
                "cleanupTasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/helper.RawMessage"
                    }
                },
                "config": {
                    "type": "object",
                    "additionalProperties": true
                },
                "configVars": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "schedule": {
                    "$ref": "#/definitions/types.TestSchedule"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/helper.RawMessage"
                    }
                },
                "timeout": {
                    "type": "string"
                }
            }
        },
        "web_api.PostTestsRegisterResponse": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "name": {
                    "type": "string"
                },
                "test_id": {
                    "type": "string"
                }
            }
        },
        "web_api.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        }
    },
...
}